backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
anthropic==0.86.0
python-dotenv==1.1.0
pydantic==2.12.5
redis==5.2.1
sqlalchemy==2.0.41
pymysql==1.1.1
cryptography==44.0.3
bcrypt==4.3.0
pyjwt==2.10.1
python-multipart==0.0.20
